Enable absolute paths w/ exec
